refactor: simplify CLI
